glMatrix u32->uintptr_t

+++ include/NL/gl/glMatrix.h
@@ -9,7 +9,7 @@ void glMatrixOrthographicCentered(nlMatrix4& m, float width, float height, float
 void glMatrixOrthographic(nlMatrix4& m, float width, float height);
 void glSetMatrix(unsigned long matrix, const nlMatrix4& m);
 void glGetMatrix(unsigned long matrix, nlMatrix4& m);
-u32 glAllocMatrix();
+uintptr_t glAllocMatrix();
 unsigned long glGetIdentityMatrix();
 void gl_MatrixStartup();
